requirements.txt
openai==2.29.0
pandas==3.0.1
pandas_ta==0.4.71b0
playwright==1.58.0
python-dotenv==1.2.2
Requests==2.32.5
simplejson==3.20.2
yfinance==1.2.0
